+++ crates/raven/src/command/mod.rs
@@ -1,4 +1,5 @@
 use clap::Subcommand;
+use raven_database::Context;
 
 mod cli;
 
@@ -10,9 +11,9 @@ pub enum RavenCmd {
 }
 
 impl RavenCmd {
-    pub fn run(self) {
+    pub fn run(self, context:&mut Context) {
         match self {
-            Self::Cli(cli) => cli.run(),
+            Self::Cli(cli) => cli.run(context),
         }
     }
 }